feat(04-function-types): Add function overload code and documentation
A new file 08.function - overload.ts was created to demonstrate function overload in TypeScript. This includes overload and implementation signatures. Also, the README.md was updated to document the concept of function overloads, providing code examples. This was done to showcase how functions can be called in different ways in TypeScript.