feat(05-object-oriented): Add abstract class example and related documentation
Added a new TypeScript file '07.abstract.ts' that demonstrates the use of abstract classes
for calculating areas of different shapes. Also updated README.md to explain the concept
of abstract classes and provide code examples. This was done to enhance understanding
of object - oriented concepts in the 05 - object - oriented module.