feat(axios): Improve axios requests and add type interfaces
This commit enhances axios requests in the application. New type interfaces for post data and random joke data are added. Requests for posts and random jokes are updated to use these interfaces. A change is made in the request function to handle types better. Also, the global response interceptor now returns response.data instead of the full response, to simplify data handling.